clean up testing
