+++ integration_test/tests/direct_cluster_tests.rs
@@ -1,10 +1,9 @@
 mod common;
 
-mod mongosqltranslate_tests {
+mod direct_cluster_tests {
     use crate::common::{
-        allocate_env, connect_with_conn_string, default_setup_connect_and_alloc_stmt,
-        disconnect_and_close_handles, fetch_and_get_data, get_column_attributes,
-        get_sql_diagnostics,
+        default_setup_connect_and_alloc_stmt, disconnect_and_close_handles, fetch_and_get_data,
+        get_column_attributes, get_sql_diagnostics,
     };
     use cstr::WideChar;
     use definitions::{
@@ -15,22 +14,7 @@ mod mongosqltranslate_tests {
     use std::ptr;
 
     #[test]
-    fn test_srv_style_uri_connection() {
-        let env_handle = allocate_env(AttrOdbcVersion::SQL_OV_ODBC3);
-        let conn_str =
-            crate::common::generate_srv_style_connection_string(Some("test".to_string()));
-        let result = connect_with_conn_string(env_handle, Some(conn_str), true);
-
-        assert!(
-            result.is_ok(),
-            "Expected successful connection, got error: {result:?}"
-        );
-
-        let _ = unsafe { Box::from_raw(env_handle) };
-    }
-
-    #[test]
-    fn test_sql_prepare_and_sql_execute_with_library_loaded_and_valid_query_and_valid_schemas_created(
+    fn test_sql_prepare_and_sql_execute_with_valid_query_and_valid_schemas_created_in_direct_cluster_mode(
     ) {
         let (env_handle, dbc, stmt) = default_setup_connect_and_alloc_stmt(
             AttrOdbcVersion::SQL_OV_ODBC3,
@@ -77,7 +61,7 @@ mod mongosqltranslate_tests {
     }
 
     #[test]
-    fn test_sql_execute_direct_with_library_loaded_and_valid_query_and_valid_schemas_created() {
+    fn test_sql_execute_direct_with_valid_query_and_valid_schemas_created_in_direct_cluster_mode() {
         let (env_handle, dbc, stmt) = default_setup_connect_and_alloc_stmt(
             AttrOdbcVersion::SQL_OV_ODBC3,
             Some(crate::common::generate_srv_style_connection_string(Some(
@@ -116,7 +100,7 @@ mod mongosqltranslate_tests {
     }
 
     #[test]
-    fn test_enterprise_mode_with_invalid_query_and_valid_schemas_created() {
+    fn test_invalid_query_with_valid_schemas_created_in_direct_cluster_mode() {
         let (env_handle, dbc, stmt) = default_setup_connect_and_alloc_stmt(
             AttrOdbcVersion::SQL_OV_ODBC3,
             Some(crate::common::generate_srv_style_connection_string(Some(
@@ -148,7 +132,7 @@ mod mongosqltranslate_tests {
     }
 
     #[test]
-    fn test_enterprise_mode_with_valid_query_and_no_sql_schemas_collection() {
+    fn test_valid_query_with_no_schemas_created_in_direct_cluster_mode() {
         let (env_handle, dbc, stmt) = default_setup_connect_and_alloc_stmt(
             AttrOdbcVersion::SQL_OV_ODBC3,
             Some(crate::common::generate_srv_style_connection_string(Some(
@@ -179,7 +163,7 @@ mod mongosqltranslate_tests {
     }
 
     #[test]
-    fn test_sql_columnsw_with_library_loaded_and_valid_schemas_created() {
+    fn test_sql_columnsw_with_valid_schemas_created_in_direct_cluster_mode() {
         let (env_handle, dbc, stmt) = default_setup_connect_and_alloc_stmt(
             AttrOdbcVersion::SQL_OV_ODBC3,
             Some(crate::common::generate_srv_style_connection_string(Some(
